% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataset.R
\name{checkDataset}
\alias{checkDataset}
\title{Function to check the dataset and detect possible issues for its use with saemixPBPK}
\usage{
checkDataset(data)
}
\arguments{
\item{data}{a dataframe to be checked.}
}
\value{
Print result on screen the result of the check
}
\description{
Function to check the dataset and detect possible issues for its use with saemixPBPK
}
\examples{


checkDataset(data)   

}
\author{
Donato Teutonico.
}
