% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForNA.R
\name{checkForNA}
\alias{checkForNA}
\title{Function to verify the presence of NAs in dataframes}
\usage{
checkForNA(d, verbose = TRUE)
}
\arguments{
\item{d}{a dataframe to be checked.}

\item{verbose}{Logical.If \code{TRUE}, the result provided by the function is a text. If \code{verbose=TRUE}, the result will be a number indicating the number of column containing NAs. In this last case, the outcome of the function may be used by other functions.}
}
\value{
Print result on screen (\code{verbose=TRUE}) or return the number of column
}
\description{
Function to verify the presence of NAs in dataframes
}
\examples{


checkForNA(d)     # By default verbose=TRUE
checkForNA(d,verbose=FALSE)

}
\author{
Donato Teutonico.
}
