% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForNumeric.R
\name{checkForNumeric}
\alias{checkForNumeric}
\title{Function to verify which column in a dataframe are numeric}
\usage{
checkForNumeric(myD, verbose = TRUE)
}
\arguments{
\item{myD}{a dataframe to be checked.}

\item{verbose}{Logical.If \code{TRUE}, the result provided by the function is a text. If \code{verbose=TRUE}, the result will be a number indicating the number of column containing NAs. In this last case, the outcome of the function may be used by other functions.}
}
\value{
Print result on screen (\code{verbose=TRUE}) or return the number of column
}
\description{
Function to verify which column in a dataframe are numeric
}
\examples{


checkForNumeric(d)     # By default verbose=TRUE
checkForNumeric(d,verbose=FALSE)

}
\author{
Donato Teutonico.
}
