% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualsSimulation.r
\name{create_individuals}
\alias{create_individuals}
\title{Function to create individuals from a dataset with the package \code{ospsuite}}
\usage{
create_individuals(
  dataset,
  sep = ",",
  weight_unit = "kg",
  height_unit = "cm",
  age_unit = "year(s)",
  progress = TRUE,
  fromFile = FALSE,
  seed = NULL
)
}
\arguments{
\item{dataset}{a dataframe or a dataset name containing information of the individuals. The dataset should contain information for:
\itemize{
  \item species (If missing it will be assumed to be \code{Human})
  \item population (required if species is \code{Human})
  \item gender (required if species is \code{Human})
  \item weight
  \item age (required if species is \code{Human})
  \item height (required if species is \code{Human})
  }}

\item{sep}{the field separator string. Values within each row of \code{dataset} are separated by this string. This argument is passed to the function \code{read.table}}

\item{weight_unit}{the unit of the weights contained in \code{dataset}. Default is "kg".}

\item{height_unit}{the unit of the heights contained in \code{dataset}. Default is "cm".}

\item{age_unit}{the unit of the ages contained in \code{"dataset"}. Default is "year(s)".}

\item{progress}{Logical. If \code{TRUE} an output on the console will be provided showing the progress of the simulation.}

\item{fromFile}{Logical. If \code{TRUE} the function will try to read a dataset file. If \code{FALSE} (default) the function will use the dataframe \code{dataset}}

\item{seed}{seed which will be set up for the simulation process}
}
\value{
A list of individuals for \code{ospsuite}.
}
\description{
Function to create individuals from a dataset with the package \code{ospsuite}
}
\examples{

myInd <- create_individuals(dataset = d, fromFile=FALSE)

}
\author{
Donato Teutonico.
}
