% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSimulations.r
\name{createSimulations}
\alias{createSimulations}
\title{Function to create individuals from a dataset with the package \code{ospsuite}}
\usage{
createSimulations(
  data,
  pkml.file,
  parameter = NULL,
  individuals = NULL,
  timeFactor = 1,
  pkmlFromFile = FALSE,
  individual.values = NULL,
  dose = NULL,
  indexTimeRemove = NULL,
  index = NULL,
  id = NULL,
  times = NULL,
  clearTimes = TRUE
)
}
\arguments{
\item{data}{a dataframe or a dataset name containing information of the individuals.}

\item{pkml.file}{link to the pkml file}

\item{parameter}{vector of parameters to estimate}

\item{individuals}{list of individuals. If missing the individual in the simulation file will be used for all simulations}

\item{timeFactor}{factor to modify the time units for unit conversion in minutes}

\item{individual.values}{specific parameter values that can be set up in the simulation}

\item{dose}{dose value}

\item{indexTimeRemove}{index of time to be removed (it can be set up to 1 to remove the time=0 simulated by default by PK-Sim if not present in the datset)}

\item{index}{vector of ID index that should be simulated. If NULL all IDs in the dataset will be simulated}

\item{id}{vector of ID that should be simulated. If NULL all IDs in the dataset will be simulated}

\item{times}{vector of time points that should be added in the simulation.If \code{clearTimes=TRUE} this vector of time points will replace the times of the simulation; otherwise it will be added to the times already simulated. 
If \code{NULL} the times to be included in the simulation will be the ones contained in \code{data}}

\item{clearTimes}{vector of ID that should be simulated. If NULL all IDs in the dataset will be simulated}
}
\value{
A list of individuals for \code{ospsuite}.
}
\description{
Function to create individuals from a dataset with the package \code{ospsuite}
}
\examples{

myInd <- create_individuals(dataset = d, fromFile=FALSE)

}
\author{
Donato Teutonico.
}
