% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_creategrid.R
\name{gridfuncd}
\alias{gridfuncd}
\title{Function to create the grid used by saemixPBPK.}
\usage{
gridfuncd(
  simulations = NULL,
  data = NULL,
  output = NULL,
  psi0 = NULL,
  psi.min = NULL,
  psi.max = NULL,
  e.obj = 0.05,
  h.obj = 0.25,
  rh = 0.9,
  nw0 = 11,
  alpha = 0.5,
  N0 = 6,
  model = modelBatch,
  seed = 12345,
  index = NULL,
  id = NULL
)
}
\arguments{
\item{simulations}{a list of simulations.}

\item{data}{a data frame with the data.}

\item{output}{a list with fields
\itemize{
  \item \code{name} path of the output
  \item \code{scaleFactor} ratio between the unit of data and the unit of model output (default=1)
  }}

\item{psi0}{a vector of initial values.}

\item{psi.min}{a vector of minimum parameter values (default = 0.05*psi0).}

\item{psi.max}{a vector of maximum parameter values (default = 10*psi0).}

\item{e.obj}{the maximum relative error between interpolations and model predictions (default=0.05).}

\item{h.obj}{the minimum RMSE between model predictions and observations (default=0.25). Used to eliminate unrealistic values in the grid.}

\item{rh}{a coefficient between 0 and 1 (default = 0.9).}

\item{nw0}{number of points in the initial grid (default = 11).}

\item{alpha}{power used to define the initial grid (default = 0.5).}

\item{N0}{number of id's used to define the initial grid (default = 6).}

\item{model}{model function (default = modelBatch).}

\item{seed}{seed used to randomly select N0 id's for the initial grid (default = 12345).}

\item{index}{a vector of selected indexes (default = NULL).}

\item{id}{a vector of selected id's (default = NULL)}
}
\value{
A list of d grids.
}
\description{
Function to create the grid used by saemixPBPK.
}
\examples{

x.grid05 <- gridfuncd(simulations=theod1.simulations, data=theo.data, psi0=c(Cl=0.005))
x.grid02 <- gridfuncd(simulations=theod2.simulations, data=theo.data, psi0=c(Cl=0.005, lipo=1), nw0=7, N.select=5, e.obj=0.02)

}
\author{
Marc Lavielle.
}
