% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_interpeval.R
\name{interpeval}
\alias{interpeval}
\title{Function to evaluate the accuracy of the interpolation.}
\usage{
interpeval(
  fit,
  rgrid,
  simulations,
  output = NULL,
  index = NULL,
  id = NULL,
  extreme = T
)
}
\arguments{
\item{fit}{output of saemix.}

\item{rgrid}{individual grids and predictions.}

\item{simulations}{a list of simulations.}

\item{output}{a list with fields
\itemize{
  \item \code{name} path of the output
  \item \code{scaleFactor} ratio between the unit of data and the unit of model output (default=1)
  }}

\item{index}{a vector of selected indexes (default: all the indexes are used).}

\item{id}{a vector of selected id's (default: all the id's are used)}

\item{extreme}{a boolean, compare the predictions obtained with extreme parameter values (default = TRUE)}
}
\value{
A list.
}
\description{
Function to evaluate the accuracy of the interpolation.
}
\examples{

r <- interpeval(fit.saemix, rgrid, simulations)

plot(r$plot[[1]])  
plot(r$plot[[2]])
plot(r$plot[[3]])
head(r$pred.extreme)

r <- interpeval(fit.saemix, rgrid, simulations, index=c(3,6,9))
r <- interpeval(fit.saemix, rgrid, simulations, id=c("id-03","id-06","id-09"))

}
\author{
Marc Lavielle.
}
