% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixModelPBPK.R
\name{modelBatch}
\alias{modelBatch}
\title{Function to run simulation batches with given parameter values}
\usage{
modelBatch(psi, simulation, output = NULL)
}
\arguments{
\item{psi}{a vector or a matrix with a number of elements, resp. with the number of columns, equal to the number d of
parameters in the model}

\item{simulation}{a list with fields
\itemize{
  \item \code{simBatch} a SimulationBatch
  \item \code{indexTimeRemove} indexes of times to remove (optional)
  }}

\item{output}{a list with fields (optional)
\itemize{
  \item \code{name} path of the output
  \item \code{scaleFactor} ratio between the unit of data and the unit of model output (default=1)
  }}
}
\value{
A vector of output values.
}
\description{
Function to run simulation batches with given parameter values
}
\author{
Marc Lavielle
}
