% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixModelPBPK.R
\name{saemixModelPBPK}
\alias{saemixModelPBPK}
\title{Function to create a SaemixModelPBPK object}
\usage{
saemixModelPBPK(
  psi0 = NULL,
  transform.par = NULL,
  x.grid = NULL,
  model = modelBatch,
  simulations = NULL,
  output = NULL,
  fixed.estim = NULL,
  omega.init = NULL,
  covariance.model = NULL,
  covariate.model = NULL,
  error.model = "combined",
  error.init = NULL,
  file = NULL,
  model.interpolate = TRUE
)
}
\arguments{
\item{psi0}{a matrix with a number of columns equal to the number d of
parameters in the model, and one (when no covariates are available) or two
(when covariates enter the model) giving the initial estimates for the fixed
effects. The column names of the matrix should be the names of the
parameters in the model, and will be used in the plots and the summaries.
When only the estimates of the mean parameters are given, psi0 may be a
named vector.}

\item{transform.par}{the distribution for each parameter (0=normal,
1=log-normal, 2=probit, 3=logit). Defaults to a vector of 1s (all parameters
have a log-normal distribution)}

\item{x.grid}{a list of d grids.}

\item{model}{name of the function used to compute the structural model (default=modelBatch).}

\item{simulations}{a list of simulations.}

\item{output}{a list with fields
\itemize{
  \item \code{name} path of the output
  \item \code{scaleFactor} ratio between the unit of data and the unit of model output (default=1)
  }}

\item{fixed.estim}{whether parameters should be estimated (1) or fixed to
their initial estimate (0). Defaults to a vector of 1s}

\item{omega.init}{a square matrix of size equal to the number of parameters
in the model, giving the initial estimate for the variance-covariance matrix
of the model.}

\item{covariance.model}{a square matrix of size equal to the number of
parameters in the model, giving the variance-covariance matrix of the model:
1s correspond to estimated variances (in the diagonal) or covariances
(off-diagonal elements). Defaults to the identity matrix}

\item{covariate.model}{a matrix giving the covariate model. Defaults to no
covariate in the model}

\item{error.model}{type of residual error model. Valid types are constant,
proportional, combined and exponential (default=combined).}

\item{error.init}{a vector of size 2 giving the initial value of a and b in
the error model. Defaults to 1 for each estimated parameter in the error
model}

\item{file}{file name with a list rgrid  that contains individual predictions previously computed).}

\item{model.interpolate}{a boolean, whether interpolation should be used or not (default=T).}
}
\value{
A SaemixModelPBPK object.
}
\description{
Function to create a SaemixModelPBPK object
}
\examples{

output <- list(name = "Organism|PeripheralVenousBlood|Theophylline|Plasma (Peripheral Venous Blood)",
               scaleFactor = 180.17/1000)


saemix.modeld <- saemixModelPBPKd(psi0=p0, 
                                  x.grid=x.grid05, 
                                  simulations=theophylline.simulations, 
                                  output=output)

}
\author{
Marc Lavielle'
}
